Some changes in SRC file

+++ src/main.c
@@ -1,15 +1,21 @@
 #include <stdio.h>
+#include <stdlib.h>
 #include "deposit.h"
 
 
 int main (){
 
-float cash=0; float day=0;
+int  cash=0; int day=0;
 printf ("Vvedite summu vklada :");
-scanf ("%f",&cash);
+scanf ("%d",&cash);
 printf ("Vvedite kol-vo dney:");
-scanf ("%f",&day);
+scanf ("%d",&day);
 check (day,cash);
 task ( day, cash);
+
+
+
+
+
 return 0;
 }
